Ventas avance 1.0

+++ Dominio/PersonalLogic.cs
@@ -83,7 +83,7 @@ namespace Dominio
 
             return Consulta.ToList();
         }
-        public void CrearArticulos(string Art, int Pre,int imp, string med ,string claSAT,string alm,string ser, byte[] Foto)
+        public void CrearArticulos(string Art, decimal Pre,decimal imp, string med ,string claSAT,string alm,string ser, byte[] Foto)
         {
             L_Database.CrearArticulo(Art, Pre,imp , med, claSAT, alm, ser, Foto);
             L_Database.SubmitChanges();
